+++ saleor/graphql/order/tests/mutations/test_order_update.py
@@ -13,7 +13,7 @@ from ....tests.utils import assert_no_permission, get_graphql_content
 
 ORDER_UPDATE_MUTATION = """
     mutation orderUpdate(
-        $id: ID!, $email: String, $address: AddressInput, $externalReference: String
+        $id: ID!, $email: String, $address: AddressInput, $externalReference: String, $privateMetadata: [MetadataInput!], $metadata: [MetadataInput!]
     ) {
         orderUpdate(
             id: $id,
@@ -21,7 +21,9 @@ ORDER_UPDATE_MUTATION = """
                 userEmail: $email,
                 externalReference: $externalReference,
                 shippingAddress: $address,
-                billingAddress: $address
+                billingAddress: $address,
+                metadata: $metadata,
+                privateMetadata: $privateMetadata,
                 }
             ) {
             errors {
@@ -31,6 +33,8 @@ ORDER_UPDATE_MUTATION = """
             order {
                 userEmail
                 externalReference
+                metadata {key, value}
+                privateMetadata {key, value}
             }
         }
     }
@@ -593,3 +597,150 @@ def test_order_update_triggers_webhooks(
     )
 
     assert wrapped_call_order_event.called
+
+
+@patch("saleor.plugins.manager.PluginsManager.order_updated")
+def test_order_update_only_metadata(
+    order_updated_webhook_mock,
+    staff_api_client,
+    permission_group_manage_orders,
+    order_with_lines,
+    graphql_address_data,
+):
+    # given
+    permission_group_manage_orders.user_set.add(staff_api_client.user)
+    order = order_with_lines
+    order_with_lines.metadata = {}
+    order_with_lines.private_metadata = {}
+    order.save()
+
+    order_id = graphene.Node.to_global_id("Order", order.id)
+
+    variables = {
+        "id": order_id,
+        "metadata": [{"key": "meta key", "value": "meta value"}],
+    }
+
+    # when
+    response = staff_api_client.post_graphql(ORDER_UPDATE_MUTATION, variables)
+    content = get_graphql_content(response)
+
+    # then
+    assert not content["data"]["orderUpdate"]["errors"]
+
+    order.refresh_from_db()
+
+    assert order.metadata == {"meta key": "meta value"}
+
+    order_updated_webhook_mock.assert_called_once_with(order, webhooks=set())
+
+
+@patch("saleor.plugins.manager.PluginsManager.order_updated")
+def test_order_update_only_private_metadata(
+    order_updated_webhook_mock,
+    staff_api_client,
+    permission_group_manage_orders,
+    order_with_lines,
+    graphql_address_data,
+):
+    # given
+    permission_group_manage_orders.user_set.add(staff_api_client.user)
+    order = order_with_lines
+    order_with_lines.metadata = {}
+    order_with_lines.private_metadata = {}
+    order.save()
+
+    order_id = graphene.Node.to_global_id("Order", order.id)
+
+    variables = {
+        "id": order_id,
+        "privateMetadata": [{"key": "meta key", "value": "meta value"}],
+    }
+
+    # when
+    response = staff_api_client.post_graphql(ORDER_UPDATE_MUTATION, variables)
+    content = get_graphql_content(response)
+
+    # then
+    assert not content["data"]["orderUpdate"]["errors"]
+
+    order.refresh_from_db()
+
+    assert order.private_metadata == {"meta key": "meta value"}
+
+    order_updated_webhook_mock.assert_called_once_with(order, webhooks=set())
+
+
+@patch("saleor.plugins.manager.PluginsManager.order_updated")
+def test_order_update_public_and_private_metadata(
+    order_updated_webhook_mock,
+    staff_api_client,
+    permission_group_manage_orders,
+    order_with_lines,
+    graphql_address_data,
+):
+    # given
+    permission_group_manage_orders.user_set.add(staff_api_client.user)
+    order = order_with_lines
+    order_with_lines.metadata = {}
+    order_with_lines.private_metadata = {}
+    order.save()
+
+    order_id = graphene.Node.to_global_id("Order", order.id)
+
+    variables = {
+        "id": order_id,
+        "privateMetadata": [{"key": "meta key", "value": "meta value"}],
+        "metadata": [{"key": "meta key", "value": "meta value"}],
+    }
+
+    # when
+    response = staff_api_client.post_graphql(ORDER_UPDATE_MUTATION, variables)
+    content = get_graphql_content(response)
+
+    # then
+    assert not content["data"]["orderUpdate"]["errors"]
+
+    order.refresh_from_db()
+
+    assert order.metadata == {"meta key": "meta value"}
+    assert order.private_metadata == {"meta key": "meta value"}
+
+    order_updated_webhook_mock.assert_called_once_with(order, webhooks=set())
+
+
+@patch("saleor.plugins.manager.PluginsManager.order_updated")
+def test_order_update_invalid_metadata(
+    order_updated_webhook_mock,
+    staff_api_client,
+    permission_group_manage_orders,
+    order_with_lines,
+    graphql_address_data,
+):
+    # given
+    permission_group_manage_orders.user_set.add(staff_api_client.user)
+    order = order_with_lines
+    order_with_lines.metadata = {}
+    order_with_lines.private_metadata = {}
+    order.save()
+
+    order_id = graphene.Node.to_global_id("Order", order.id)
+
+    variables = {
+        "id": order_id,
+        # Empty key is invalid
+        "metadata": [{"key": "", "value": "meta value"}],
+    }
+
+    # when
+    response = staff_api_client.post_graphql(ORDER_UPDATE_MUTATION, variables)
+    content = get_graphql_content(response)
+
+    # then
+    errors = content["data"]["orderUpdate"]["errors"]
+
+    assert errors[0]["field"] == "metadata"
+    assert errors[0]["code"] == "REQUIRED"
+
+    order.refresh_from_db()
+    assert order.metadata == {}